requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.55.3
httpx==0.27.0
chromadb==0.5.5
faiss-cpu==1.8.0
pypdf2==3.0.1
tiktoken==0.7.0
streamlit==1.38.0
python-dotenv==1.0.1
pysqlite3-binary==0.5.3 
